Split __init__.py into modules
